optimize LruDiskCache and LruMemoryCache

+++ sketch/src/main/java/me/xiaopan/sketch/cache/LruDiskCache.java
@@ -35,6 +35,7 @@ import me.xiaopan.sketch.util.DiskLruCache;
 import me.xiaopan.sketch.util.NoSpaceException;
 import me.xiaopan.sketch.util.SketchUtils;
 import me.xiaopan.sketch.util.UnableCreateDirException;
+import me.xiaopan.sketch.util.UnableCreateFileException;
 
 public class LruDiskCache implements DiskCache {
     protected String logName = "LruDiskCache";
@@ -46,23 +47,33 @@ public class LruDiskCache implements DiskCache {
     private DiskLruCache cache;
     private Configuration configuration;
     private Map<String, ReentrantLock> editLockMap;
+    private boolean closed;
 
     public LruDiskCache(Context context, Configuration configuration, int appVersionCode, int maxSize) {
         this.context = context;
         this.maxSize = maxSize;
         this.appVersionCode = appVersionCode;
         this.configuration = configuration;
+        this.cacheDir = SketchUtils.getSketchCacheDir(context, true, DISK_CACHE_DIR_NAME);
+    }
+
+    private boolean checkCache(boolean checkCacheDir) {
+        if (checkCacheDir) {
+            return cache != null && !cache.isClosed() && cacheDir != null && cacheDir.exists();
+        } else {
+            return cache != null && !cache.isClosed();
+        }
     }
 
     /**
-     * 安装磁盘缓存，当缓存目录不存在的时候回再次安装
+     * 安装磁盘缓存
      */
-    private synchronized void installDiskCache(boolean force) {
-        // 好好的就不安装了
-        if (!force && cache != null && cacheDir != null && cacheDir.exists()) {
+    private synchronized void reinstallDiskCache() {
+        if (closed) {
             return;
         }
 
+        // 旧的要关闭
         if (cache != null) {
             try {
                 cache.close();
@@ -72,24 +83,23 @@ public class LruDiskCache implements DiskCache {
             cache = null;
         }
 
+        // 重置缓存目录
+        cacheDir = SketchUtils.getSketchCacheDir(context, true, DISK_CACHE_DIR_NAME);
+
+        // 创建缓存目录，然后检查空间并创建个文件测试一下
         try {
-            cacheDir = SketchUtils.getCacheDir(context, DISK_CACHE_DIR_NAME, true, DISK_CACHE_RESERVED_SPACE_SIZE, true, true, 10);
+            cacheDir = SketchUtils.buildCacheDir(context, cacheDir, DISK_CACHE_RESERVED_SPACE_SIZE, true, true, 10);
         } catch (NoSpaceException e) {
             e.printStackTrace();
-            cacheDir = e.dir;
-
-            if (configuration.getErrorCallback() != null) {
-                configuration.getErrorCallback().onInstallDiskCacheFailed(e, cacheDir);
-            }
+            configuration.getErrorCallback().onInstallDiskCacheFailed(e, cacheDir);
             return;
-        }
-
-        if (!cacheDir.exists() && !cacheDir.mkdirs()) {
-            UnableCreateDirException exception = new UnableCreateDirException("Unable to create dir: " + cacheDir.getPath());
-            exception.printStackTrace();
-            if (configuration.getErrorCallback() != null) {
-                configuration.getErrorCallback().onInstallDiskCacheFailed(exception, cacheDir);
-            }
+        } catch (UnableCreateDirException e) {
+            e.printStackTrace();
+            configuration.getErrorCallback().onInstallDiskCacheFailed(e, cacheDir);
+            return;
+        } catch (UnableCreateFileException e) {
+            e.printStackTrace();
+            configuration.getErrorCallback().onInstallDiskCacheFailed(e, cacheDir);
             return;
         }
 
@@ -97,18 +107,19 @@ public class LruDiskCache implements DiskCache {
             cache = DiskLruCache.open(cacheDir, appVersionCode, 1, maxSize);
         } catch (IOException e) {
             e.printStackTrace();
-
-            if (configuration.getErrorCallback() != null) {
-                configuration.getErrorCallback().onInstallDiskCacheFailed(e, cacheDir);
-            }
+            configuration.getErrorCallback().onInstallDiskCacheFailed(e, cacheDir);
         }
     }
 
     @Override
     public boolean exist(String uri) {
+        if (closed) {
+            return false;
+        }
+
         // 这里有些特殊，只有当没有尝试安装过的时候才会尝试安装，这是由于目前此方法只在helper中用到，而Helper对性能要求较高
-        if (cacheDir == null) {
-            installDiskCache(false);
+        if (!checkCache(false)) {
+            reinstallDiskCache();
         }
 
         return cache != null && cache.exist(uriToDiskCacheKey(uri));
@@ -116,7 +127,13 @@ public class LruDiskCache implements DiskCache {
 
     @Override
     public Entry get(String uri) {
-        installDiskCache(false);
+        if (closed) {
+            return null;
+        }
+
+        if (!checkCache(true)) {
+            reinstallDiskCache();
+        }
 
         DiskLruCache.SimpleSnapshot snapshot = null;
         try {
@@ -129,7 +146,13 @@ public class LruDiskCache implements DiskCache {
 
     @Override
     public Editor edit(String uri) {
-        installDiskCache(false);
+        if (closed) {
+            return null;
+        }
+
+        if (!checkCache(true)) {
+            reinstallDiskCache();
+        }
 
         DiskLruCache.Editor diskEditor = null;
         try {
@@ -137,7 +160,7 @@ public class LruDiskCache implements DiskCache {
         } catch (IOException e) {
             e.printStackTrace();
             // 发生异常的时候（比如SD卡被拔出，导致不能使用），尝试重建DiskLruCache，能显著提高遇错恢复能力
-            installDiskCache(true);
+            reinstallDiskCache();
             try {
                 diskEditor = cache != null ? cache.edit(uriToDiskCacheKey(uri)) : null;
             } catch (IOException e1) {
@@ -149,7 +172,6 @@ public class LruDiskCache implements DiskCache {
 
     @Override
     public File getCacheDir() {
-        installDiskCache(false);
         return cacheDir;
     }
 
@@ -160,9 +182,10 @@ public class LruDiskCache implements DiskCache {
 
     @Override
     public String uriToDiskCacheKey(String uri) {
-        if (SketchUtils.checkSuffix(uri, ".apk")) {
-            uri += ".icon";
-        }
+        // 由于DiskLruCache会在文件名后面加序列号，因此这里不用再处理了
+//        if (SketchUtils.checkSuffix(uri, ".apk")) {
+//            uri += ".icon";
+//        }
         try {
             return URLEncoder.encode(uri, "UTF-8");
         } catch (UnsupportedEncodingException e) {
@@ -173,34 +196,65 @@ public class LruDiskCache implements DiskCache {
 
     @Override
     public long getSize() {
+        if (closed) {
+            return 0;
+        }
+
         return cache != null ? cache.size() : 0;
     }
 
     @Override
     public void clear() {
+        if (closed) {
+            return;
+        }
+
         if (cache != null) {
             try {
                 cache.delete();
             } catch (IOException e) {
                 e.printStackTrace();
             }
+            cache = null;
         }
-        installDiskCache(true);
+
+        reinstallDiskCache();
+    }
+
+    @Override
+    public boolean isClosed() {
+        return closed;
     }
 
     @Override
     public void close() {
+        if (closed) {
+            return;
+        }
+
+        closed = true;
+
         if (cache != null) {
             try {
                 cache.close();
             } catch (IOException e) {
                 e.printStackTrace();
             }
+            cache = null;
+        }
+
+        if (editLockMap != null) {
+            editLockMap.clear();
+            editLockMap = null;
         }
     }
 
     @Override
     public synchronized ReentrantLock getEditLock(String key) {
+        if (closed) {
+            return null;
+        }
+
         if (key == null) {
             return null;
         }
@@ -231,6 +285,8 @@ public class LruDiskCache implements DiskCache {
                 .append("maxSize").append("=").append(Formatter.formatFileSize(context, maxSize))
                 .append(",")
                 .append("appVersionCode").append("=").append(appVersionCode)
+                .append(",")
+                .append("cacheDir").append("=").append(cacheDir.getPath())
                 .append(")");
     }
